`janus_test_util`: generic `dummy_vdaf` module
I needed a VDAF with a non-unit aggregation parameter for testing
collect job logic. While extending the existing `fake::Vdaf` to do that,
I decided to move it into a module under `janus_test_util`, since
`aggregator.rs` is already too big.
