+++ test_util/src/lib.rs
@@ -11,6 +11,8 @@ use rand::{thread_rng, Rng};
 use ring::aead::{LessSafeKey, UnboundKey, AES_128_GCM};
 use std::sync::{Arc, Mutex};
 
+pub mod dummy_vdaf;
+
 /// The Janus database schema.
 pub static SCHEMA: &str = include_str!("../../db/schema.sql");
